refactor(test): extract ChromeDeviceTest superclass and cleanup
